requirements.txt
discord.py
flask
openai